REMAINING SOME TESTING || FOLDER COMPRESSION

+++ src/com/company/Node.java
@@ -2,10 +2,10 @@ package com.company;
 
 public class Node implements  Comparable<Node>{
     final char character;
-    final short frequency;
+    final int frequency;
     final Node leftChild,rightChild;
 
-    Node(char character, short frequency, Node leftChild, Node rightChild) {
+    Node(char character, int frequency, Node leftChild, Node rightChild) {
         this.character = character;
         this.frequency = frequency;
         this.leftChild=leftChild;
@@ -18,7 +18,7 @@ public class Node implements  Comparable<Node>{
 
     @Override
     public int compareTo(final Node o) {
-        final int result=Short.compare(this.frequency,o.frequency);
+        final int result=Integer.compare(this.frequency,o.frequency);
 
         return result!=0? result:Integer.compare(this.character,o.character);
     }
